requirements.txt
streamlit
borsapy
google-genai
pandas
plotly
tabulate